Rename .storybook folder to have no dots in the prefix.
