+++ .github/workflows/storybook.yml
@@ -12,7 +12,7 @@ on:
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
-      - '.storybook/**'
+      - 'storybook/**'
       - '**.scss'
       - '.nvmrc'
       - '**/package.json'
@@ -29,7 +29,7 @@ on:
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
-      - '.storybook/**'
+      - 'storybook/**'
       - '**.scss'
       - '.nvmrc'
       - '**/package.json'
