+++ .github/workflows/visual-regression.yml
@@ -15,7 +15,7 @@ on:
       - '!assets/**/*.test.js'
       - '**.stories.js'
       - 'tests/backstop/**'
-      - '.storybook/**'
+      - 'storybook/**'
       - '**.scss'
       - '.nvmrc'
       - '**/package.json'
