+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
 	"search.exclude": {
-		".storybook/data": true,
+		"storybook/data": true,
 		"**/dist": true,
 		"**/node_modules/": true,
 		"**/composer.lock": true,
