+++ package.json
@@ -11,7 +11,7 @@
 		"build:production": "webpack --mode=production",
 		"build:test": "npm run build:production -- --include-tests",
 		"prebuild:storybook": "npm run remove-dist",
-		"build:storybook": "build-storybook -c .storybook -o dist --quiet --loglevel warn",
+		"build:storybook": "build-storybook -c storybook -o dist --quiet --loglevel warn",
 		"remove-dist": "rimraf ./dist",
 		"prezip": "composer run autoload-includes",
 		"zip": "./bin/release",
@@ -29,19 +29,19 @@
 		"test:analyze": "npm run build:production -- --env.ANALYZE",
 		"test:js": "cross-env BABEL_ENV=test NODE_ENV=test jest --config=tests/js/jest.config.js --passWithNoTests",
 		"test:js:watch": "npm run test:js -- --watch",
-		"test:storybook": "npm run test:js -- --testMatch '<rootDir>/.storybook/*.test.js'",
+		"test:storybook": "npm run test:js -- --testMatch '<rootDir>/storybook/*.test.js'",
 		"test:e2e": "cross-env NODE_OPTIONS=--experimental-abortcontroller npm run test:e2e -w ./tests/e2e",
 		"test:e2e:interactive": "cross-env NODE_OPTIONS=--experimental-abortcontroller npm run test:e2e:interactive -w ./tests/e2e",
 		"test:e2e:watch": "cross-env NODE_OPTIONS=--experimental-abortcontroller npm run test:e2e:watch -w ./tests/e2e",
 		"test:e2e:watch:interactive": "cross-env NODE_OPTIONS=--experimental-abortcontroller npm run test:e2e:watch:interactive -w ./tests/e2e",
-		"storybook": "start-storybook -p 9001 -c .storybook --no-version-updates",
+		"storybook": "start-storybook -p 9001 -c storybook --no-version-updates",
 		"storybook:vrt": "npm run storybook -- --ci --quiet",
 		"backstopjs": "./bin/backstop",
 		"lint": "npm run lint:js && npm run lint:css",
 		"lint:js": "eslint --no-eslintrc --config .eslintrc.json --ignore-path .eslintignore --ext js,jsx,ts,tsx .",
 		"lint:js-fix": "prettier --write --config .prettierrc.js --ignore-path .prettierignore \"**/*.{js,jsx,ts,tsx,.yml,.yaml}\"",
-		"lint:css": "stylelint ./assets/sass ./.storybook/assets/sass",
-		"lint:css-fix": "stylelint ./assets/sass ./.storybook/assets/sass --fix",
+		"lint:css": "stylelint ./assets/sass ./storybook/assets/sass",
+		"lint:css-fix": "stylelint ./assets/sass ./storybook/assets/sass --fix",
 		"env:start": "npm run env:start -w ./tests/e2e",
 		"env:stop": "npm run env:stop -w ./tests/e2e",
 		"env:reset-site": "npm run env:reset-site -w ./tests/e2e",
@@ -69,7 +69,7 @@
 		"assets/sass/**/*.scss": [
 			"stylelint"
 		],
-		".storybook/assets/sass/**/*.scss": [
+		"storybook/assets/sass/**/*.scss": [
 			"stylelint"
 		],
 		"*.php": [
