+++ tests/backstop/scenarios.js
@@ -34,10 +34,10 @@ const path = require( 'path' );
 /**
  * Internal dependencies
  */
-const storybookConfig = require( '../../.storybook/main' );
+const storybookConfig = require( '../../storybook/main' );
 const rootURL = 'file:///src/dist/iframe.html?id=';
 
-const storybookDir = path.resolve( __dirname, '../../.storybook' );
+const storybookDir = path.resolve( __dirname, '../../storybook' );
 const storyFiles = flatten(
 	storybookConfig.stories
 		.map( ( storiesPattern ) =>
